GraphAdjust/requirements.txt
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.6
chardet==4.0.0
chex==0.0.8
click==8.0.1
colorama==0.4.4
configparser==5.0.2
Cython==0.29.22
dataclasses==0.6
debugpy==1.3.0
decorator==4.4.2
defusedxml==0.7.1
Django==3.1.4
flatbuffers==1.12
flax==0.3.4
frozendict==1.2
future==0.18.2
gast==0.3.3
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
gin-config==0.4.0
gitdb==4.0.7
GitPython==3.1.18
google-auth==1.33.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.14.1
grpcio==1.32.0
h5py==2.10.0
idna==2.10
image-geometry==1.13.0
imageio==2.6.1
imageio-ffmpeg==0.4.2
importlib-metadata==3.10.1
interactive-markers==1.11.5
ipykernel==6.0.3
ipython==7.25.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
jax==0.2.17
jaxlib==0.1.68+cuda111
jedi==0.18.0
Jinja2==3.0.1
joblib==0.16.0
joint-state-publisher==1.12.15
jsonschema==3.2.0
jupyter-client==6.1.12
jupyter-core==4.7.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.0
kdl-parser-py==1.13.1
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
Klampt==0.8.7
kornia==0.5.0
laser-geometry==1.6.7
llvmlite==0.34.0
lpips-pytorch===latest
Markdown==3.3.4
MarkupSafe==2.0.1
mathutils==2.81.2
matplotlib==3.4.3
matplotlib-inline==0.1.2
message-filters==1.14.11
mistune==0.8.4
moveit-commander==1.0.8
moveit-core==1.0.8
moveit-ros-planning-interface==1.0.8
moveit-ros-visualization==1.0.8
msgpack==1.0.2
nbclient==0.5.3
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.5
notebook==6.4.0
numba==0.51.0
numpy==1.21.1
oauthlib==3.1.1
olefile==0.46
open3d==0.10.0.0
opencv-python==4.2.0.32
opt-einsum==3.3.0
optax==0.0.9
packaging==21.0
pandas==0.25.3
pandocfilters==1.4.3
parso==0.8.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
plyfile==0.7.1
portalocker==2.0.0
posix-ipc==1.0.5
prometheus-client==0.11.0
promise==2.3
prompt-toolkit==3.0.19
protobuf==3.14.0
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybullet==2.8.7
pycparser==2.20
pygifsicle==1.0.4
pyglet==1.5.16
PyGLM==2.5.1
Pygments==2.9.0
PyOpenGL==3.1.5
pyparsing==2.4.7
PyQt5==5.15.5
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.0
pyquaternion==0.9.5
pyrsistent==0.18.0
python-dateutil==2.8.1
python-fcl==0.0.12
python-qt-binding==0.4.4
pytorch-ignite==0.2.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
regex==2021.7.6
requests==2.25.1
requests-oauthlib==1.3.0
resource-retriever==1.12.6
Rtree==0.9.7
rviz==1.13.18
sacrebleu==1.5.1
scikit-image==0.16.2
scikit-learn==0.24.2
scipy==1.4.1
Send2Trash==1.7.1
sensor-msgs==1.12.8
sentry-sdk==1.2.0
Shapely==1.7.1
shortuuid==1.0.1
six==1.15.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
smmap==4.0.0
sqlparse==0.4.1
srdfdom==0.5.2
subprocess32==3.5.4
sysv-ipc==1.1.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
tensorflow-hub==0.11.0
termcolor==1.1.0
terminado==0.10.1
testpath==0.5.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
threadpoolctl==2.1.0
tifffile==2021.7.2
toolz==0.11.1
topic-tools==1.14.11
tornado==6.1
tqdm==4.43.0
trac-ik-python==1.5.1
traitlets==5.0.5
trimesh==3.9.10
typing-extensions==3.7.4.3
urdfdom-py==0.4.5
urllib3==1.25.11
wandb==0.10.33
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
xacro==1.13.12
zipp==3.5.0
